Default selected/ not selected
Hi, I was using this library in one of my quiz app.
Now the issue is by default first option is selected . But which is not normal for a quiz app so there should be an option for default selection. if user puts the value for default selected during call then only it will be default selected else should not. I have made some changes  in CustomRadioButton(
......
    this.currentSelected,
....

) 

.....
final int currentSelected;



and on init state


@override
  void initState() {
    super.initState();
    if(widget.currentSelected!=null){
      currentSelectedLabel = widget.buttonLables[widget.currentSelected];

    }

  }

when you call 


CustomRadioButton(
........
            enableShape: true,
            hight: 40,
            width: 60,
            currentSelected: 3,
//or dont give anything if you dont want

......